Move render-type resolution tests into realm-server
resolveRenderType is pure runtime-common logic tested only on the realm server,
so the tests live directly in packages/realm-server/tests as a self-contained
qunit module importing from @cardstack/runtime-common, rather than via the
runtime-common/tests + consumer SharedTests indirection (which is reserved for
suites that run in both host and realm-server).
